requirements.txt
lida
lida[transformers]
torch
streamlit
pymongo
python-dotenv
fastapi[standard]
redis